requirements.txt
anyascii==0.3.1
arrow==0.15.8
asgiref==3.5.2
astroid==2.11.6
autopep8==1.5.7
beautifulsoup4==4.9.3
black==22.6.0
boto3==1.24.57
botocore==1.27.57
braintree==4.16.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
Collectfast==2.2.0
colorama==0.4.5
confusable-homoglyphs==3.2.0
cryptography==36.0.2
defusedxml==0.7.1
dill==0.3.5.1
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==4.0.8
django-crispy-forms==1.14.0
django-etc==1.4.0
django-filter==21.1
django-flatpickr==1.0.1
django-hitcount==1.3.5
django-modelcluster==6.0
django-mpesa==2.0.14
django-on-heroku==1.1.2
django-permissionedforms==0.1
django-registration==3.3
django-storages==1.13.1
django-taggit==2.1.0
django-timezone-field==5.0
django-treebeard==4.5.1
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
flake8==3.9.2
gunicorn==20.1.0
html5lib==1.1
httplib2==0.9.2
idna==3.3
importlib-resources==5.10.0
isort==5.10.1
jeepney==0.8.0
jmespath==1.0.1
keras==2.8.0
keyring==23.8.2
l18n==2021.3
lazy-object-proxy==1.7.1
Markdown==3.3.7
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.22.4
oauthlib==3.2.1
openpyxl==3.0.10
packaging==20.9
pandas==1.4.3
pathspec==0.9.0
Pillow==9.0.1
pkginfo==1.8.3
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
PyJWT==2.3.0
pylint==2.14.4
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
SecretStorage==3.3.3
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
telepath==0.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
tqdm==4.64.0
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.9
wagtail==3.0.1
wagtail-color-panel==1.4.0
wagtailemoji==1.0.2
wagtailfontawesome==1.2.1
wagtailmedia==0.9.0
waitress==2.1.2
webencodings==0.5.1
whitenoise==5.3.0
Willow==1.4.1
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
zipp==3.8.1
